+++ src/index.test.ts
@@ -31,6 +31,8 @@ import {
   getResourceInfoWithAcl,
   getPodOwner,
   isPodOwner,
+  getLinkedResourceUrlAll,
+  getEffectiveAccess,
   isContainer,
   isRawData,
   getContentType,
@@ -179,6 +181,8 @@ it("exports the public API from the entry file", () => {
   expect(getResourceInfoWithAcl).toBeDefined();
   expect(getPodOwner).toBeDefined();
   expect(isPodOwner).toBeDefined();
+  expect(getLinkedResourceUrlAll).toBeDefined();
+  expect(getEffectiveAccess).toBeDefined();
   expect(isContainer).toBeDefined();
   expect(isRawData).toBeDefined();
   expect(getContentType).toBeDefined();