+++ src/index.ts
@@ -28,6 +28,8 @@ export {
   getResourceInfo,
   getPodOwner,
   isPodOwner,
+  getLinkedResourceUrlAll,
+  getEffectiveAccess,
   FetchError,
 } from "./resource/resource";
 export {